[Gold IV] Title: 알파벳, Time: 1572 ms, Memory: 2024 KB -BaekjoonHub

+++ 백준/Gold/1987. 알파벳/알파벳.cc
@@ -1,6 +1,6 @@
 #include<bits/stdc++.h>
 using namespace std;
-using ll = long long;
+using ll = int;
 using pll = pair<ll,ll>;
 using tlll = tuple<ll,ll,ll>;
 int main(){
